Trend Dataframe
We created functions to get trends for monthly income and spend and yearly income and spend and put them into a pandas data frame. We also created a function to get the differences in income throughout a month and put it into a pandas data frame. We created functions to remove data from our database. We worked on this through live share on Vscode with Justin, Hamiz and Divyam.